refactor(room): get list, update, delete api
