+++ src/hospital/hospital.service.ts
@@ -334,113 +334,6 @@ export class HospitalService {
     return wards;
   }
 
-  async getRoomList(hospitalId: string) {
-    const wards = await this.getWardList(hospitalId);
-    const result = [];
-
-    for (const ward of wards) {
-      const wardId = ward.ward_id;
-
-      const rooms = await this.roomRepository
-        .createQueryBuilder('room')
-        .select('room.id')
-        .addSelect('room.roomNumber')
-        .addSelect('room.createdAt')
-        .addSelect('room.currentPatient')
-        .addSelect('room.icuCheck')
-        .leftJoin('room.ward', 'ward')
-        .where('ward.id = :wardId', { wardId })
-        .execute();
-
-      for (const room of rooms) {
-        const createdAt_temp = room.room_createdAt.toISOString().split('T')[0];
-        const createdAt_temp2 = createdAt_temp.split('-');
-
-        result.push({
-          ward_id: ward.ward_id,
-          ward_name: ward.ward_name,
-          room_id: room.room_id,
-          room_createdAt:
-            createdAt_temp2[0].substring(2) +
-            '/' +
-            createdAt_temp2[1] +
-            '/' +
-            createdAt_temp2[2],
-          room_number: room.room_roomNumber,
-          room_currentPatient: room.room_currentPatient,
-          room_icuCheck: room.room_icuCheck,
-        });
-      }
-    }
-    return result;
-  }
-
-  async updateRoom(hospitalId: string, requestDto: UpdateRoomDto) {
-    const hospital = await this.findHospital(hospitalId);
-
-    const { id, ...updateData } = requestDto;
-    const room = await this.roomRepository.findOne({
-      where: { id: id },
-      relations: { ward: true },
-    });
-
-    if (!room) {
-      throw new BadRequestException({
-        statusCode: HttpStatus.BAD_REQUEST,
-        message: ['존재하지 않는 병실입니다.'],
-        error: 'BAD_REQUEST',
-      });
-    }
-
-    const ward = await this.findWard(hospital.id, room.ward.name);
-
-    if (ward.length != 1) {
-      throw new BadRequestException({
-        statusCode: HttpStatus.BAD_REQUEST,
-        message: ['올바르지 않은 병실 정보입니다.'],
-        error: 'BAD_REQUEST',
-      });
-    }
-
-    const result = await this.roomRepository.update({ id }, updateData);
-
-    if (result.affected > 0) {
-      return requestDto;
-    }
-  }
-
-  async deleteRoom(hospitalId: string, requestDto: DeleteRoomDto) {
-    const hospital = await this.findHospital(hospitalId);
-    const room = await this.roomRepository.findOne({
-      where: { id: requestDto.id },
-      relations: { ward: true },
-    });
-
-    if (!room) {
-      throw new BadRequestException({
-        statusCode: HttpStatus.BAD_REQUEST,
-        message: ['존재하지 않는 병실입니다.'],
-        error: 'BAD_REQUEST',
-      });
-    }
-
-    const ward = await this.findWard(hospital.id, room.ward.name);
-
-    if (ward.length != 1) {
-      throw new BadRequestException({
-        statusCode: HttpStatus.BAD_REQUEST,
-        message: ['올바르지 않은 병실 정보입니다.'],
-        error: 'BAD_REQUEST',
-      });
-    }
-
-    const result = await this.roomRepository.delete({ id: requestDto.id });
-
-    if (result.affected > 0) {
-      return 'success';
-    }
-  }
-
   async updatePatient(requestDto: UpdatePatientDto, hospitalId: string) {
     const hospital = await this.findHospital(hospitalId);
