+++ src/room/dto/request-dto/update-room.dto.ts
@@ -1,11 +1,22 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsBoolean, IsNotEmpty, IsNumber, IsOptional } from 'class-validator';
+import {
+  IsBoolean,
+  IsNotEmpty,
+  IsNumber,
+  IsOptional,
+  IsString,
+} from 'class-validator';
 
 export class UpdateRoomDto {
   @IsNumber()
   @IsNotEmpty()
   @ApiProperty({ description: '수정을 원하는 병실 아이디', example: 3 })
   id: number;
+
+  @IsNotEmpty()
+  @IsString()
+  @ApiProperty({ description: '병실이 포함된 병동 이름', example: '201동' })
+  wardName: string;
 
   @IsOptional()
   @IsNumber()
