+++ src/room/entities/room.entity.ts
@@ -1,7 +1,8 @@
 import { Common } from '../../entities/common.entity';
-import { Entity, Column, ManyToOne, OneToMany } from 'typeorm';
+import { Entity, Column, ManyToOne, OneToMany, JoinColumn } from 'typeorm';
 import { Patient } from '../../hospital/entities/patient.entity';
 import { Ward } from '../../ward/entities/ward.entity';
+import { Hospital } from 'src/hospital/entities/hospital.entity';
 
 @Entity()
 export class Room extends Common {
@@ -18,8 +19,15 @@ export class Room extends Common {
   icuCheck: boolean;
 
   @ManyToOne(() => Ward, (ward) => ward.rooms, { onDelete: 'CASCADE' })
+  @JoinColumn()
   ward: Ward;
 
   @OneToMany(() => Patient, (patient) => patient.room)
   patients: Patient[];
+
+  @ManyToOne(() => Hospital, (hospital) => hospital.rooms, {
+    onDelete: 'CASCADE',
+  })
+  @JoinColumn()
+  hospital: Hospital;
 }
